% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/YJtransformation.R
\name{DYJtrans}
\alias{DYJtrans}
\title{Derivative of Yeo-Johnson transformation}
\usage{
DYJtrans(y, theta)
}
\arguments{
\item{y}{Input numeric value}

\item{theta}{Parameter of YJ transformation}
}
\value{
derivative of YJ transform at specific value
}
\description{
Derivative of Yeo-Johnson transformation
}
\examples{
dtrans <- DYJtrans(2,0.5)
}
