% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/YJtransformation.R
\name{IYJtrans}
\alias{IYJtrans}
\title{Inverse of Yeo-Johnson transformation}
\usage{
IYJtrans(y, theta)
}
\arguments{
\item{y}{Numeric value}

\item{theta}{parameter for the inverse transformation}
}
\value{
Inverse transformed value
}
\description{
It computes the inverse of Yeo and Johnson (2000) transformation.
             See the paper by Yeo and Johnson (2000) for details.
}
\examples{
Itrans = IYJtrans(-2,0.5)
}
