% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/YJtransformation.R
\name{slog}
\alias{slog}
\title{Logarithmic transformation}
\usage{
slog(y)
}
\arguments{
\item{y}{The original value}
}
\value{
Log transformed value
}
\description{
Logarithmic transformation
}
\examples{
ylog = slog(5)
}
