% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/YJtransformation.R
\name{spower}
\alias{spower}
\title{Power transformation}
\usage{
spower(y, pw)
}
\arguments{
\item{y}{Input value}

\item{pw}{Parameter of power transformation}
}
\value{
Power transformed value
}
\description{
Power transformation
}
\examples{
ypow = spower(2,0.5)
}
