% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/YJtransformation.R
\name{YJtrans}
\alias{YJtrans}
\title{Yeo-Johnson transformation}
\usage{
YJtrans(y, theta)
}
\arguments{
\item{y}{numeric variable}

\item{theta}{parameter of the transformation}
}
\value{
Transformed value
}
\description{
Yeo-Johnson transformation is used to transform a continuous (numeric) variable
             so that the resulting variable looks more normally distributed.
             See the details in the paper by
             Yeo, I. K., and Johnson, R. A. (2000). A new family of power transformations to improve normality or symmetry, Biometrika,87, 954--959.
}
\examples{
ytrans <-YJtrans(5,0.2)
ytrans <-YJtrans(-5,0.5)
}
